modified the saveReviewQuestions function to allow displayPriority to be set. (before, displayPriority was ignored)
Now, if displayPriority is non-zero then it is set.  The reason we
do not set it when it is zero is that in the UI call to this function
(which does not set displayPriority), displayPriority is initialized
to zero.  It would be better if that code unset displayPriority and then
this function was modified to check whether displayPriority is set, rather
than non-zero.

With the caveat above, displayPriority can be effectively set with the API
using non-zero priorities.